Separate member/invitations team views
